+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to the z/OS FTP Plug-in for Zowe CLI will be documented in this file.
 
+## Recent Changes
+- Add codePage setting in profile.
+
 ## `2.0.2`
 - Refine help of partitioned dataset allocation.